pyproject.toml
[project]
name = "myvelocity-ai-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "subramanya-dev",email = "connect.subbu08@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0.0"
dependencies = [
    "fastapi (>=0.136.0,<0.137.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "langchain (>=1.2.15,<2.0.0)",
    "weaviate-client (>=4.20.5,<5.0.0)",
    "anthropic (>=0.96.0,<0.97.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-anthropic (>=1.4.1,<2.0.0)",
    "sentence-transformers (>=5.4.1,<6.0.0)",
    "langchain-weaviate (>=0.0.6,<0.0.7)"
]

[tool.poetry]
packages = [{include = "myvelocity_ai_chatbot", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
myvelocitybot = "src.myvelocity_ai_chatbot.main:app"
